refactor : 1주차 피드백 반영
- 도메인 서비스의 불필요한 추상화 제거
